refactor: point 엔티티 위치 변경에 따른 수정
