+++ src/main/java/com/example/humorie/mypage/repository/PointRepository.java
@@ -1,7 +1,7 @@
 package com.example.humorie.mypage.repository;
 
 import com.example.humorie.account.entity.AccountDetail;
-import com.example.humorie.mypage.entity.Point;
+import com.example.humorie.account.entity.Point;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
